Added GameObject and rect classes.
The GameObject is responsible for handling all logic tied to a single
entity in the game world. The rect class is just a representation of a
rectangle with some helper functions to access corners and to check if
two rectangles intersect. The rect class is spelled with a lower case r
to mimick the GLM library.