Integrated texture loading into Lua
You can now load a texture from Lua. Also added support for listing all
the files in a directory from Lua. For now you can only ask for a table
of filenames in a specific directory. In the future you will get a table
of file and directory objects so that you can choose to traverse the
filesystem deeper if you would like to.