MacOSX mostly working.
The solution compiles and runs but there are a few problems, most
notably the shaders are not supported on this machine. Figure out if
version 330 works on both macbook and desktop win machine. Also, try to
figure out a nicer way to include SDL and glew/opengl. Right now we
have to do #ifdef’s everywhere to make sure we include it right.
Instead we should make an include header that does this for us that we
can in turn include.