Updated Input class
Changed from checking SDL_Keycode to checking SDL_Scancode. Keycode is
for the character that a key produces whilst Scancode is the actual
physical key. So even though the characters WASD may be mapped to
different locations on different keyboards, the scancode will always be
the same. This is why it's better for games to check scancodes instead
of keycodes.